refactor(message-service): 채팅방 이전 대화 내역 조회 MessageService로 이동 #34
